+++ api/post-video.php
@@ -5,6 +5,9 @@
  * 
  * Read from a form post call an video file and a text file containing the jwt.
  * Test jwt if it comes from a valid user and change their video to the given one.
+ * 
+ * NOTE: change PHP.ini file at "upload_max_filesize" to accept files 55MB or less.
+ * 		 change PHP.ini file at "post_max_size" to accept post files 55MB or less.
  */
 
 // required headers
@@ -73,7 +76,9 @@ if(($_POST["jwt"])) {
 
 					// here delete the previous video if they have any
 					if(!empty($user->video_path)) {
-						unlink("./uploads/".$user->video_path);
+						if(file_exists("./uploads/".$user->video_path)) {
+							unlink("./uploads/".$user->video_path);
+						}
 					}
 
 					// give user the new path to video
